+++ src/Col/styles/index.less
@@ -1,14 +1,2 @@
-@import '../../styles/common';
-
-// Columns
-//
-// Common styles for small and large grid columns
-.rs-col {
-  float: inline-start;
-  position: relative;
-  // Prevent columns from collapsing when empty
-  min-height: 1px;
-  // Inner gutter via padding
-  padding-inline-start: calc(var(--rs-grid-gutter) / 2);
-  padding-inline-end: calc(var(--rs-grid-gutter) / 2);
-}
+// Col styles are included in Grid component
+// --------------------------------------------------
