+++ src/Grid/index.tsx
@@ -1,8 +1,12 @@
 import Grid from './Grid';
+import Row from './Row';
+import Col from './Col';
 
 // export types
 export type { GridProps } from './Grid';
+export type { RowProps } from './Row';
+export type { ColProps } from './Col';
 
 // export components
-export { Grid };
+export { Grid, Row, Col };
 export default Grid;
