+++ src/Grid/styles/columns.less
@@ -1,0 +1,12 @@
+// Columns
+//
+// Common styles for small and large grid columns
+.rs-col {
+  float: inline-start;
+  position: relative;
+  // Prevent columns from collapsing when empty
+  min-height: 1px;
+  // Inner gutter via padding
+  padding-inline-start: calc(var(--rs-grid-gutter) / 2);
+  padding-inline-end: calc(var(--rs-grid-gutter) / 2);
+}
