+++ src/Grid/styles/index.less
@@ -2,6 +2,8 @@
 @import './mixin.less';
 @import './grid-framework.less';
 @import './variables.less';
+@import './row.less';
+@import './columns.less';
 
 //
 // Grid system
