+++ src/Row/index.tsx
@@ -1,7 +1,7 @@
-import Row from './Row';
+import Row from '../Grid/Row';
 
 // export types
-export type { RowProps } from './Row';
+export type { RowProps } from '../Grid/Row';
 
 // export components
 export { Row };
