+++ src/Row/styles/index.less
@@ -1,57 +1,2 @@
-@import '../../styles/common';
-@import '../../Grid/styles/mixin';
-@import '../../Grid/styles/grid-framework';
-
-// Row
-//
-// Rows contain and clear the floats of your columns.
-.rs-row {
-  display: flex;
-  flex-flow: row wrap;
-  margin-inline-start: calc(var(--rs-grid-gutter) / -2);
-  margin-inline-end: calc(var(--rs-grid-gutter) / -2);
-
-  // At the top of the alignment.
-  &-top {
-    align-items: flex-start;
-  }
-
-  // Align center.
-  &-middle {
-    align-items: center;
-  }
-
-  // At the bottom of the alignment.
-  &-bottom {
-    align-items: flex-end;
-  }
-
-  //##==== Layout (X-Axis).
-
-  // Pack items from the start.
-  &-start {
-    justify-content: flex-start;
-  }
-
-  // Pack items around the center.
-  &-center {
-    justify-content: center;
-  }
-
-  // Pack items from the end.
-  &-end {
-    justify-content: flex-end;
-  }
-
-  // Distribute items evenly.
-  // The first item is flush with the start, the last is flush with the end.
-  &-space-between {
-    justify-content: space-between;
-  }
-
-  // Distribute items evenly.
-  // Items have a half-size space on either end.
-  &-space-around {
-    justify-content: space-around;
-  }
-}
+// Col styles are included in Grid component
+// --------------------------------------------------
